Made space for the new brain code.
By moving the old common code into legacy module. And make it available
on the top level so that i don't need to change old code.
